+++ tests/runTests.R
@@ -1,0 +1,3 @@
+require("ATACseqQC") || stop("unable to load Package:ATACseqQC")
+require("testthat") || stop("unable to load testthat")
+test_check("ATACseqQC")